Extract flow execution logic into separate runFlow function
